backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pyjwt[crypto]
passlib[bcrypt]
pymongo
python-dotenv